refactor: move string converter to dedicated file
